Allow overriding the HTML widget in GtkhtmlEditor.
This adds a construct-only "html" property which allows a custom GtkHTML
(or subclass) to be used in a GtkhtmlEditor instance.

Evolution uses this to provide an EWebView instance, which implements
more policies than a plain GtkHTML instance.

Also remove some obsolete initialization logic from the core GtkHTML
widget which was meant for the old Bonobo-based editor component.