Bug 672582 - Fix selection behavior inside <iframe>
Since kinetic scrolling was introduced in Gtk+, GtkScrolledWindow
has had its own GdkWindow. This means we need to use the allocation
of the GtkScrolledWindow rather than its child GtkHTML widget when we
want to translate pointer coordinates relative to the iframe, because
a widget's GtkAllocation is relative to the parent GdkWindow.