Paragraph style drawn incorrectly after font style change
For example, if there is a Header 1 paragraph, and it whole is selected,
and then a bold/italic/underline is changed, then the drawn paragraph
style is like Normal, even the above chooser still shows Header 1. Writing
a new character makes the text drawing fixed, showing the text correctly.

The issue was that HTMLObject's parent change didn't mark the object
properties for recalculation, thus there was used stale style set, calculated
in time when there was no parent set on the object.