htmlobject: handle containers with non-cursor-able children
If the cursor is right at the beginning of a table like

    <table>
      <tr>
	<td>
	  <a name="">
	  </a>
	</td>
	<td>
	  foo
	</td>
      </tr>
    </table>
    bar

then move_object_cursor() would find the first TableCell (the first
<td>), and its Anchor.  The anchor contains nothing, so it can't take
the cursor.

However, move_object_cursor() would jump to the step where it does,
"find me the next object within the current object's parent", but for
the entire Table, not for the TableCell.  Thus it would miss the second
<td> that actually contains a "foo" string, and it would skip past the
Table to its next cursor-able sibling, the "bar" string.